I updated my exercises

+++ Exercises/EX008.java
@@ -5,20 +5,20 @@ public class EX008{
     Scanner entrada = new Scanner(System.in);
     // Lê os valores de a, b, c, d, e e f
     System.out.print("Digite o valor de a: ");
-    double a = entrada.nextDouble();
+    int a = entrada.nextInt();
     System.out.print("Digite o valor de b: ");
-    double b = entrada.nextDouble();
+    int b = entrada.nextInt();
     System.out.print("Digite o valor de c: ");
-    double c = entrada.nextDouble();
+    int c = entrada.nextInt();
     System.out.print("Digite o valor de d: ");
-    double d = entrada.nextDouble();
+    int d = entrada.nextInt();
     System.out.print("Digite o valor de e: ");
-    double e = entrada.nextDouble();
+    int e = entrada.nextInt();
     System.out.print("Digite o valor de f: ");
-    double f = entrada.nextDouble();
+    int f = entrada.nextInt();
     // Calcula os valores de x e y
-    double x = (c * e - b * f) / (a * e - b * d);
-    double y = (a * f - c * d) / (a * e - b * d);
+    int x = (c * e - b * f) / (a * e - b * d);                                      
+    int y = (a * f - c * d) / (a * e - b * d);
     // Exibe o resultado
     System.out.println("O valor de x é: " + x);
     System.out.println("O valor de y é: " + y);
